+++ src/spring.h
@@ -22,8 +22,11 @@ void compress(const std::string &temp_dir, const string_list &input_paths,
               const bool pairing_only_flag, const bool no_quality_flag,
               const bool no_ids_flag, const string_list &quality_options,
               const int compression_level, const std::string &note,
-                            const log_level verbosity_level = log_level::info,
-                            const bool audit_flag = false);
+              const log_level verbosity_level = log_level::info,
+              const bool audit_flag = false,
+              const std::string &r3_path = std::string(),
+              const std::string &i1_path = std::string(),
+              const std::string &i2_path = std::string());
 
 void decompress(const std::string &temp_dir, const string_list &input_paths,
                 const string_list &output_paths, const int num_thr,